Remove deprecated enable / disable repos feature
