+++ CHANGELOG.md
@@ -19,6 +19,8 @@ All notable changes to `src-cli` are documented in this file.
 
 ### Removed
 
+- The `src repos enable|disable` commands were removed as they are no longer supported.
+
 ## 3.28.3
 
 ### Fixed
